update a lot

+++ src/main/java/com/alan10607/leaf/model/ContLike.java
@@ -19,7 +19,7 @@ public class ContLike {
     private String id;
 
     @Id
-    private int no;
+    private Long no;
 
     @Id
     private String userId;
